requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

# HTTP provider for API calls
apache-airflow-providers-http==4.5.0

# PostgreSQL provider for database operations
apache-airflow-providers-postgres==5.7.0

# BigQuery provider for Google Cloud BigQuery operations - updated to compatible version
apache-airflow-providers-google==10.8.0

# AWS provider for S3/MinIO and Glue operations - compatible version
apache-airflow-providers-amazon==8.7.0

# boto3 for direct S3/MinIO operations
boto3==1.34.0

# SQL Server provider for Microsoft SQL Server operations
apache-airflow-providers-microsoft-mssql==4.3.0

# Parquet file processing
pyarrow==14.0.1
pandas==2.1.4

# PySpark for ETL processing
pyspark==3.5.4

# Additional dependencies for PySpark with S3/MinIO support
boto3==1.34.0
botocore==1.34.0

# PostgreSQL adapter - using psycopg (newer, more reliable)
# psycopg[binary]==3.1.18  # Commented out - using Airflow's built-in provider instead

# SQLAlchemy for database operations (compatible with Airflow providers)
sqlalchemy==1.4.53

# psycopg2 for PostgreSQL connections
psycopg2-binary==2.9.9